finished home and input
