+++ src/css/createbutton.css
@@ -1,0 +1,19 @@
+
+.create-button {
+    background-color: #9c27b0 ;
+    display: inline;
+    border: none;
+    outline: none;
+    padding: 1em 2em;
+    color: white;
+    border-radius: 3px;
+    box-shadow: 2em;
+    font-weight: 500;
+    margin-top: 2em;
+    margin-bottom: 2em;
+    transition: background-color 0.3s; /* ホバー時のトランジション */
+}
+  
+.create-button:hover {
+    background-color: #b638d9; /* 少し明るい色に変更 */
+}
