+++ src/css/header.css
@@ -1,4 +1,4 @@
-*{
+body{
     box-sizing: border-box;
     margin: 0;
     padding: 0;
@@ -25,7 +25,7 @@
 
 .user {
     display: flex;
-    gap: 20px;
+    gap: 15px;
     align-items: center;
     color: aliceblue;
 }
@@ -35,4 +35,7 @@
     margin: 0 auto;
     border-top: 1px solid rgb(109, 109, 136);
     opacity: 2;
+}
+.userimg{
+    width: 2em;
 }
