+++ src/css/home.css
@@ -20,8 +20,8 @@
     display: flex;
     flex-direction: column;
     align-items: flex-start;
+    justify-content: center;
     padding-left: 2em;
-    gap: 2em;
 }
 
 .urw{
@@ -36,11 +36,13 @@
 .cp{
     color: white;
     font-size: 3em;
+    margin-block-start: 0.2em !important;
+    margin-block-end: 0.2em !important;
 }
 
 .bcp{
     color: rgb(223, 223, 223);
-    text-align: left;
+    margin-bottom: 1.8em;
 }
 
 .wbtn{
